'Functioning' but I had to disable MsgPool_deinit.
MsgPool_deinit is hanging in the loop waiting for all of the msgs to
be returned. I'm not sure why that is when it 'works' if the ring buffer
isn't being used.

I'm also thinking maybe the MsgPool should just use MpscRingBuffer since
it is a finite resource as its used now. Although I could see in the
future it might be useful to add to it incrementally.

Also, performance is terrible when using 'test'.