Update README.md and bottom line this is slow.
As coded, it's slow and more complex than I was hoping. It could
definitely be made faster and maybe there is a simpler way to do it
and achieve my goal of higher performance. But for now I'm going to
have sadie use test-mpscfifo2 type algorithm which gives decent
performance and is easier to use than test-mpscfifo for large messages
and implement an unbounded fifo.